Move CHighPrecisionTimer to util
